CR:Kanaiya rename Plugin Name
