+++ android/build.gradle
@@ -1,4 +1,4 @@
-group 'com.atritechnocrat.flutter_picker'
+group 'com.atritechnocrat.flutter_media_picker'
 version '1.0-SNAPSHOT'
 
 buildscript {
