+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.atritechnocrat.flutter_picker">
+  package="com.atritechnocrat.flutter_media_picker">
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
 </manifest>
