+++ android/src/main/kotlin/com/atritechnocrat/flutter_media_picker/FileFetcher.kt
@@ -1,4 +1,4 @@
-package com.atritechnocrat.flutter_picker
+package com.atritechnocrat.flutter_media_picker
 
 import android.annotation.TargetApi
 import android.content.ContentUris
