+++ example/android/app/build.gradle
@@ -38,7 +38,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.atritechnocrat.flutter_picker_example"
+        applicationId "com.atritechnocrat.flutter_media_picker_example"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
