+++ example/lib/MultiSelectorModel.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_picker/data/MediaFile.dart';
+import 'package:flutter_media_picker/data/MediaFile.dart';
 
 class MultiSelectorModel extends ChangeNotifier {
   Set<MediaFile> _selectedItems = Set();
