+++ pubspec.yaml
@@ -1,8 +1,8 @@
-name: flutter_picker
-description: A new Flutter project.
+name: flutter_media_picker
+description: Flutter plugin for selecting media file from the Android and iOS image library.
 version: 0.0.1
-author:
-homepage:
+author: Atri Technocrat <developers@atritechnocrat.com>
+homepage: http://www.atritechnocrat.com
 
 environment:
   sdk: ">=2.1.0 <3.0.0"
@@ -25,8 +25,8 @@ flutter:
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
-    androidPackage: com.atritechnocrat.flutter_picker
-    pluginClass: FlutterPickerPlugin
+    androidPackage: com.atritechnocrat.flutter_media_picker
+    pluginClass: FlutterMediaPickerPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
