+++ test/flutter_picker_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_picker/flutter_picker.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('flutter_picker');
